Move the content of clinica/workflows to adequate location
